cast worker to int

+++ app.py
@@ -44,7 +44,7 @@ def hire(queue=None):
                 except KeyError:
                     i = 0
                 if i == 0:
-                    workers = math.ceil(queue.count/15.0)
+                    workers = (int)math.ceil(queue.count/15.0)
                     #heroku_app.processes['worker'].scale(workers)
                     cloud._http_resource(method='POST',
                                      resource=('apps', 'sudokusolver', 'ps', 'scale'),
